move logger to separate mod
